Final Attempt Before Deadline
We attempted to use keras and an LSTM neural network to classify our price predictions. It had very similar results to our previous attempts. Since we were able to achieve 55% using our confidence metrics. We are going to present with that data, and try and improve upon the model after due date.